+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 Please choose versions by [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+- feat: add Scaler interface with counterfeiter directive and ParseScalerDefault helper
+
 ## v0.4.13
 
 - update Go to 1.26.6 and update dependencies